REF Merge TournamentHeader and TournamentPlanning in Tournament
